move Card component into its own folder
