+++ src/components/Card/Card.module.css
@@ -1,0 +1,36 @@
+.card {
+  background: white;
+  border-radius: 20px;
+  padding: 30px;
+  box-shadow: 0 10px 30px rgba(0, 0, 0, 0.2);
+  border: 1px solid rgba(255, 255, 255, 0.1);
+  transition: all 0.3s ease;
+}
+
+.card:hover {
+  box-shadow: 0 15px 40px rgba(0, 0, 0, 0.3);
+  transform: translateY(-2px);
+}
+
+.cardLarge {
+  padding: 40px;
+}
+
+.cardSmall {
+  padding: 20px;
+}
+
+@media (max-width: 768px) {
+  .card {
+    padding: 20px;
+    border-radius: 16px;
+  }
+  
+  .cardLarge {
+    padding: 30px;
+  }
+  
+  .cardSmall {
+    padding: 15px;
+  }
+}
