+++ src/pages/GamePage.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useRef, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import Header from '../components/Header';
-import Card from '../components/Card';
+import Card from '../components/Card/Card';
 import Button from '../components/Button/Button';
 import Question from '../game/Question';
 import AnswerList from '../game/AnswerList';
