+++ src/pages/ResultPage.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import Header from '../components/Header';
-import Card from '../components/Card';
+import Card from '../components/Card/Card';
 import Button from '../components/Button/Button';
 import { useResults } from '../hooks/useResults';
 import type { AnswerHistory } from '../types';
