+++ src/pages/StartPage.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import Header from '../components/Header';
 import Button from '../components/Button/Button';
-import Card from '../components/Card';
+import Card from '../components/Card/Card';
 import SettingsForm from '../components/SettingsForm';
 import Modal from '../components/Modal';
 import { useQuizData } from '../hooks/useQuizData';
