+++ src/pages/UserProfilePage.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useParams, Link } from 'react-router-dom';
 import Header from '../components/Header';
-import Card from '../components/Card';
+import Card from '../components/Card/Card';
 import Button from '../components/Button/Button';
 
 const UserProfilePage: React.FC = () => {
